refactor: replace C-style casts with static_cast/reinterpret_cast
**Summary:** Replaced legacy C-style casts with C++ style casts (`static_cast`, `reinterpret_cast`).

**Why:** This is part of the "Junior Job" to modernize the codebase and improve type safety.

**Changes:**

* Replaced `(Type)` with `static_cast<Type>` where appropriate.
* Verified the changes by compiling the project with `clangd` and running the application with **GCC** (g++ 14.2.0) and **Clang** (clang 19.1.7) on Debian 13.4.

+++ app/GUI/kcmdline.cpp
@@ -261,7 +261,7 @@ QString KCMDLine::command() const
 KrActionBase::ExecType KCMDLine::execType() const
 {
     KConfigGroup grp(krConfig, "Private");
-    int i = grp.readEntry("Command Execution Mode", (int)0);
+    int i = grp.readEntry("Command Execution Mode", 0);
     return execModesMenu[i];
 }
